+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 0.2.0 (2018-06-10)
+
+* Added feature to checkout and commit timesheets
+  directly from Sublime Text.
+  Sublime's Build System is used for this.
+
 # 0.1.0 (2018-06-02)
 
 * Initial release.